--- a/Data/Shaders/Render/text_notifications.glsl
+++ b/Data/Shaders/Render/text_notifications.glsl
@@ -10,8 +10,8 @@
 //
 
 // Inputs
-in vec4 inText1;  // Position (vec2), TextureCoordinates (vec2)
-in float inText2; // Alpha
+in vec4 inTextNotification1;  // Position (vec2), TextureCoordinates (vec2)
+in float inTextNotification2; // Alpha
 
 // Outputs
 out vec2 vertexTextureCoordinates;
@@ -19,9 +19,9 @@ out float vertexAlpha;
 
 void main()
 {
-    vertexTextureCoordinates = inText1.zw; 
-    vertexAlpha = inText2;
-    gl_Position = vec4(inText1.xy, -1.0, 1.0);
+    vertexTextureCoordinates = inTextNotification1.zw; 
+    vertexAlpha = inTextNotification2;
+    gl_Position = vec4(inTextNotification1.xy, -1.0, 1.0);
 }