--- a/FloatingSandbox/MainFrame.cpp
+++ b/FloatingSandbox/MainFrame.cpp
@@ -704,15 +704,19 @@ void MainFrame::OnPostInitializeTrigger(wxTimerEvent & /*event*/)
             ImageSize(
                 mMainGLCanvas->GetSize().GetWidth(),
                 mMainGLCanvas->GetSize().GetHeight()),
-            [this, splash]() // Allow deferred execution
+            [this, splash]() // Allow deferred execution, capturing splash dialog by value
             {
                 mMainGLCanvasContext->SetCurrent(*(splash->GetOpenGLCanvas()));
 
             },
             [this]()
             {
+                ////LogMessage("TODOTEST: Swapping buffers...");
+
                 assert(!!mMainGLCanvas);
                 mMainGLCanvas->SwapBuffers();
+
+                ////LogMessage("TODOTEST: ...buffers swapped.");
             },
             *mResourceLocator,
             [this, &splash](float progress, std::string const & message)
@@ -1072,10 +1076,10 @@ void MainFrame::OnGameTimerTrigger(wxTimerEvent & /*event*/)
         mToolController->UpdateSimulation();
 
         // Update game - will also render
-        LogMessage("TODOTEST: MainFrame::OnGameTimerTrigger: Running game iteration; IsSplashShown=",
-            !!mSplashScreenDialog ? std::to_string(mSplashScreenDialog->IsShown()) : "<NoSplash>",
-            " IsMainGLCanvasShown=", !!mMainGLCanvas ? std::to_string(mMainGLCanvas->IsShown()) : "<NoCanvas>",
-            " IsFrameShown=", std::to_string(this->IsShown()));
+        ////LogMessage("TODOTEST: MainFrame::OnGameTimerTrigger: Running game iteration; IsSplashShown=",
+        ////    !!mSplashScreenDialog ? std::to_string(mSplashScreenDialog->IsShown()) : "<NoSplash>",
+        ////    " IsMainGLCanvasShown=", !!mMainGLCanvas ? std::to_string(mMainGLCanvas->IsShown()) : "<NoCanvas>",
+        ////    " IsFrameShown=", std::to_string(this->IsShown()));
         assert(!!mGameController);
         mGameController->RunGameIteration();
 
@@ -1804,7 +1808,7 @@ void MainFrame::OnNormalScreenMenuItemSelected(wxCommandEvent & /*event*/)
 void MainFrame::OnMuteMenuItemSelected(wxCommandEvent & /*event*/)
 {
     assert(!!mUIPreferencesManager);
-    mUIPreferencesManager->SetGlobalMute(mMuteMenuItem->IsChecked());
+    mUIPreferencesManager->SetGlobalMute(mMuteMenuItem->IsChecked());    
 }
 
 void MainFrame::OnHelpMenuItemSelected(wxCommandEvent & /*event*/)