--- a/FloatingSandbox/MainFrame.h
+++ b/FloatingSandbox/MainFrame.h
@@ -291,7 +291,7 @@ private:
     {
         if (!mHasWindowBeenShown)
         {
-            LogMessage("TODOTEST: MainFrame::AfterGameRender: showing frame");
+            LogMessage("MainFrame::AfterGameRender: showing frame");
 
             this->Show(true);