--- a/FloatingSandbox/SettingsManager.cpp
+++ b/FloatingSandbox/SettingsManager.cpp
@@ -126,20 +126,20 @@ BaseSettingsManager<GameSettings>::BaseSettingsManagerFactory SettingsManager::M
     ADD_GC_SETTING(float, WaterContrast);
     ADD_GC_SETTING(float, WaterLevelOfDetail);
     ADD_GC_SETTING(bool, ShowShipThroughOcean);
-    ADD_GC_SETTING(DebugShipRenderMode, DebugShipRenderMode);
-    ADD_GC_SETTING(OceanRenderMode, OceanRenderMode);
+    ADD_GC_SETTING(DebugShipRenderModeType, DebugShipRenderMode);
+    ADD_GC_SETTING(OceanRenderModeType, OceanRenderMode);
     ADD_GC_SETTING(size_t, TextureOceanTextureIndex);
     ADD_GC_SETTING(rgbColor, DepthOceanColorStart);
     ADD_GC_SETTING(rgbColor, DepthOceanColorEnd);
     ADD_GC_SETTING(rgbColor, FlatOceanColor);
-    ADD_GC_SETTING(LandRenderMode, LandRenderMode);
+    ADD_GC_SETTING(LandRenderModeType, LandRenderMode);
     ADD_GC_SETTING(size_t, TextureLandTextureIndex);
     ADD_GC_SETTING(rgbColor, FlatLandColor);
-    ADD_GC_SETTING(VectorFieldRenderMode, VectorFieldRenderMode);
+    ADD_GC_SETTING(VectorFieldRenderModeType, VectorFieldRenderMode);
     ADD_GC_SETTING(bool, ShowShipStress);
     ADD_GC_SETTING(bool, DrawHeatOverlay);
     ADD_GC_SETTING(float, HeatOverlayTransparency);
-    ADD_GC_SETTING(ShipFlameRenderMode, ShipFlameRenderMode);
+    ADD_GC_SETTING(ShipFlameRenderModeType, ShipFlameRenderMode);
     ADD_GC_SETTING(float, ShipFlameSizeAdjustment);
 	ADD_GC_SETTING(bool, DrawHeatBlasterFlame);