Completed work to run all OpenGL code on a separate thread on multi-core boxes.
Performance results (default_ship.png benchmark):
* Dell box (single-core): UNCHANGED (i.e. 30-31FPS)
* Samsung (4-core, 1GHz): 9.4FPS -> 10.5FPS
* HP (2-core, 2GHz): 16.9FPS -> 20.7FPS